WIP: issue #1; refactoring.
